feat: improved workflows
